ISSUE-1641: move table/database/table_function defintion to catalog
