+++ query/src/datasources/local/local_database.rs
@@ -25,15 +25,15 @@ use common_planners::CreateTablePlan;
 use common_planners::DropTablePlan;
 use common_planners::TableEngineType;
 
-use crate::catalogs::impls::database_catalog::LOCAL_TBL_ID_BEGIN;
-use crate::catalogs::meta::InMemoryMetas;
-use crate::catalogs::meta::TableFunctionMeta;
-use crate::catalogs::meta::TableMeta;
+use crate::catalogs::Database;
+use crate::catalogs::InMemoryMetas;
+use crate::catalogs::TableFunctionMeta;
+use crate::catalogs::TableMeta;
+use crate::catalogs::LOCAL_TBL_ID_BEGIN;
 use crate::datasources::local::CsvTable;
 use crate::datasources::local::MemoryTable;
 use crate::datasources::local::NullTable;
 use crate::datasources::local::ParquetTable;
-use crate::datasources::Database;
 
 pub struct LocalDatabase {
     tables: RwLock<InMemoryMetas>,
@@ -43,7 +43,7 @@ pub struct LocalDatabase {
 impl LocalDatabase {
     pub fn create() -> Self {
         LocalDatabase {
-            tables: RwLock::new(InMemoryMetas::new()),
+            tables: RwLock::new(InMemoryMetas::create()),
             tbl_id_seq: AtomicU64::new(LOCAL_TBL_ID_BEGIN),
         }
     }
@@ -136,7 +136,7 @@ impl Database for LocalDatabase {
         };
 
         let mut tables = self.tables.write();
-        let table_meta = TableMeta::new(Arc::from(table), self.next_db_id());
+        let table_meta = TableMeta::create(Arc::from(table), self.next_db_id());
         tables.insert(table_meta);
         Ok(())
     }
