+++ query/src/datasources/system/contributors_table_test.rs
@@ -17,6 +17,7 @@ use common_planners::*;
 use common_runtime::tokio;
 use futures::TryStreamExt;
 
+use crate::catalogs::Table;
 use crate::datasources::system::*;
 use crate::datasources::*;
