+++ query/src/datasources/system/numbers_table_test.rs
@@ -18,6 +18,7 @@ use common_planners::*;
 use common_runtime::tokio;
 use futures::TryStreamExt;
 
+use crate::catalogs::Table;
 use crate::datasources::system::*;
 use crate::datasources::*;
