+++ query/src/datasources/system/processes_table.rs
@@ -30,7 +30,7 @@ use common_planners::Statistics;
 use common_streams::DataBlockStream;
 use common_streams::SendableDataBlockStream;
 
-use crate::datasources::Table;
+use crate::catalogs::Table;
 use crate::sessions::DatafuseQueryContextRef;
 use crate::sessions::ProcessInfo;
