+++ query/src/datasources/system/system_database.rs
@@ -22,15 +22,15 @@ use common_metatypes::MetaVersion;
 use common_planners::CreateTablePlan;
 use common_planners::DropTablePlan;
 
-use crate::catalogs::impls::database_catalog::SYS_TBL_ID_BEGIN;
-use crate::catalogs::impls::database_catalog::SYS_TBL_ID_END;
-use crate::catalogs::meta::InMemoryMetas;
-use crate::catalogs::meta::TableFunctionMeta;
-use crate::catalogs::meta::TableMeta;
+use crate::catalogs::Database;
+use crate::catalogs::InMemoryMetas;
+use crate::catalogs::Table;
+use crate::catalogs::TableFunction;
+use crate::catalogs::TableFunctionMeta;
+use crate::catalogs::TableMeta;
+use crate::catalogs::SYS_TBL_ID_BEGIN;
+use crate::catalogs::SYS_TBL_ID_END;
 use crate::datasources::system;
-use crate::datasources::Database;
-use crate::datasources::Table;
-use crate::datasources::TableFunction;
 
 pub struct SystemDatabase {
     tables: InMemoryMetas,
@@ -68,8 +68,8 @@ impl SystemDatabase {
         ];
         let tbl_meta_list = table_list
             .iter()
-            .map(|t| TableMeta::new(t.clone(), next_id()));
-        let mut tables = InMemoryMetas::new();
+            .map(|t| TableMeta::create(t.clone(), next_id()));
+        let mut tables = InMemoryMetas::create();
         for tbl in tbl_meta_list.into_iter() {
             tables.insert(tbl);
         }
@@ -85,7 +85,7 @@ impl SystemDatabase {
             let name = tbl_func.name();
             table_functions.insert(
                 name.to_string(),
-                Arc::new(TableFunctionMeta::new(
+                Arc::new(TableFunctionMeta::create(
                     tbl_func.clone(),
                     tables
                         .name2meta
