+++ query/src/datasources/system/system_factory.rs
@@ -16,8 +16,8 @@ use std::sync::Arc;
 
 use common_exception::Result;
 
+use crate::catalogs::Database;
 use crate::datasources::system::SystemDatabase;
-use crate::datasources::Database;
 
 pub struct SystemFactory;
