+++ query/src/datasources/system/tables_table.rs
@@ -25,8 +25,8 @@ use common_planners::Statistics;
 use common_streams::DataBlockStream;
 use common_streams::SendableDataBlockStream;
 
-use crate::catalogs::catalog::Catalog;
-use crate::datasources::Table;
+use crate::catalogs::Catalog;
+use crate::catalogs::Table;
 use crate::sessions::DatafuseQueryContextRef;
 
 pub struct TablesTable {
